Added ListItemBanner. Moved all banners to banner directory.
This is the mixed composable of Exercise and Workout banners.
